requirements.txt
aiomysql
python-dotenv
